[Clean] Rename `component` to `components`
