+++ src/popup/components/Header.scss
