+++ src/popup/components/Scripts.scss
